+++ src/message.h
@@ -125,5 +125,5 @@ public:
 
 bool GetTxOfMessage(const std::vector<unsigned char> &vchMessage, CTransaction& tx);
 bool BuildMessageJson(const CMessage& message, UniValue& oName, const std::string &strPrivKey="");
-int GetMessageExpiration(const CMessage& message);
+uint64_t GetMessageExpiration(const CMessage& message);
 #endif // MESSAGE_H
